chore: AddWord View-VC 분리 (#10)
